+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Renderable : new isFlippedX & isFlippedY getter to check the current renderable "flip" state
 - Renderable : fix implementation and usage of the isDirty flag across Renderables
 - Renderable : fix a regression with transformation when a renderable is directly added to a me.Container object
+- Renderable : apply a tint color to BitmapText if a `fillStyle` is specified in the constructor (WebGL mode only)
 - Renderer : new optional parameter for createCanvas() to request an OffScreenCanvas if supported by the browser
 - Renderer : use OffscreenCanvas by default for all secondary canvas (e.g. debugPanel, double buffering, font texture in WebGL)
 - Renderer : properly fallback to the Canvas Renderer including when the WebGL renderer is not able to compile the built-in shaders